skills/tracking-parallel-agent-work/tool/pyproject.toml
[project]
name = "work-extractor"
version = "0.1.0"
description = "Extract work blocks from AI agent conversation logs (JSONL)"
readme = { text = "", content-type = "text/plain" }
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Jose R. Prieto" },
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
work-extractor = "work_extractor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/work_extractor"]

[tool.hatch.build.targets.wheel.force-include]
"src/work_extractor/config.yaml" = "work_extractor/config.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
